+++ api/src/db/seeds/development/campaigns.js
@@ -1,5 +1,6 @@
 const yakbak = require('yakbak')
 const path = require('path')
+const chance = require('chance').Chance()
 const http = require('http')
 const tmWorker = require('../../../tm_clock')
 
@@ -45,7 +46,8 @@ exports.seed = async (knex) => {
           // Make all the hashtags predictable
           const campaigns = await knex('campaigns').select()
           const fixedCampaigns = campaigns.map(c => knex('campaigns').where('id', c.id).update({
-            campaign_hashtag: `project-${c.tasker_id}-${c.tm_id}`
+            campaign_hashtag: `project-${c.tasker_id}-${c.tm_id}`,
+            status: chance.pickone(['ARCHIVED', 'DRAFT', 'PUBLISHED'])
           }))
 
           await Promise.all(fixedCampaigns)