+++ api/src/routes/campaigns.js
@@ -15,6 +15,7 @@ module.exports = async (req, res) => {
   const page = req.query.page || 1
   const search = req.query.q || ''
   const tm = req.query.tm || ''
+  const includeArchived = req.query.includeArchived || false
   const complMin = req.query.compl_min || 0
   const complMax = req.query.compl_max || 100
   const validMin = req.query.valid_min || 0
@@ -22,9 +23,16 @@ module.exports = async (req, res) => {
   const sortType = req.query.sortType || 'Most Recently Created'
 
   try {
-    let query = db('campaigns').whereNotNull('campaign_hashtag').whereNot('status', 'DRAFT').join(
+    let query = db('campaigns').whereNotNull('campaign_hashtag').join(
       db('taskers').select('name as tm_name', 'id as tasker_t_id', 'type').as('t'),
       'tasker_id', 't.tasker_t_id')
+
+    if (includeArchived === 'true') {
+      // Including archived campaigns means having both PUBLISHED and ARCHIVED, no DRAFT
+      query = query.whereNot('status', 'DRAFT')
+    } else {
+      query = query.where('status', 'PUBLISHED')
+    }
 
     // Total count of campaigns that have a hashtag
     const [{ allCount }] = await query.clone().count('id as allCount')