+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Campaigns can be filtered to include or exclude archived campaigns
+
 ## [v1.3.5] - 2019-07-19
 ### Changed
 - Switch campaign logic for homepage to using recency instead of priority